+++ types/mempool/sender_nonce.go
@@ -1,9 +1,10 @@
 package mempool
 
 import (
+	crand "crypto/rand"
+	"encoding/binary"
 	"fmt"
 	"math/rand"
-	"time"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/auth/signing"
@@ -12,62 +13,50 @@ import (
 
 var (
 	_ Mempool  = (*senderNonceMempool)(nil)
-	_ Iterator = (*nonceMempoolIterator)(nil)
+	_ Iterator = (*senderNonceMepoolIterator)(nil)
 )
 
 type senderTxs struct {
-	txQueue *huandu.SkipList
-	head    *huandu.Element
+	cursor *huandu.Element
 }
 
-func newSenderTxs() senderTxs {
+func newSenderTxs(tx *huandu.Element) senderTxs {
 	return senderTxs{
-		head:    nil,
-		txQueue: huandu.New(huandu.LessThanFunc(txKeyLessNonce)),
+		cursor: tx,
 	}
 }
 
-func (s *senderTxs) insert(key txKey, tx sdk.Tx) {
-	s.txQueue.Set(key, tx)
-	s.head = s.txQueue.Front()
-}
-
-func (s *senderTxs) getMove() *huandu.Element {
-	if s.head == nil {
+func (s *senderTxs) next() *huandu.Element {
+	if s.cursor == nil {
 		return nil
 	}
-	currentHead := s.head
-	s.head = s.head.Next()
-	return currentHead
-}
-
-func (s *senderTxs) remove(key txKey) error {
-	res := s.txQueue.Remove(key)
-	if res == nil {
-		return ErrTxNotFound
-	}
-	if s.head == res {
-		s.head = s.txQueue.Front()
-	}
-	return nil
+	currentCursor := s.cursor
+	s.cursor = s.cursor.Next()
+	return currentCursor
 }
 
 type senderNonceMempool struct {
-	senders map[string]*senderTxs
+	senders map[string]*huandu.SkipList
 	rnd     *rand.Rand
 }
 
+// NewSenderNonceMempool creates a new mempool that prioritizes transactions by nonce, the lowest first.
 func NewSenderNonceMempool() Mempool {
-	senderMap := make(map[string]*senderTxs)
+	senderMap := make(map[string]*huandu.SkipList)
 	snp := &senderNonceMempool{
 		senders: senderMap,
 	}
-	snp.setSeed(time.Now().UnixNano())
+
+	var seed int64
+	binary.Read(crand.Reader, binary.BigEndian, &seed)
+	snp.setSeed(seed)
+
 	return snp
 }
 
+// NewSenderNonceMempoolWithSeed creates a new mempool that prioritizes transactions by nonce, the lowest first and sets the random seed.
 func NewSenderNonceMempoolWithSeed(seed int64) Mempool {
-	senderMap := make(map[string]*senderTxs)
+	senderMap := make(map[string]*huandu.SkipList)
 	snp := &senderNonceMempool{
 		senders: senderMap,
 	}
@@ -80,6 +69,8 @@ func (snm *senderNonceMempool) setSeed(seed int64) {
 	snm.rnd = rand.New(s1)
 }
 
+// Insert adds a tx to the mempool. It returns an error if the tx does not have at least one signer.
+// priority is ignored.
 func (snm *senderNonceMempool) Insert(_ sdk.Context, tx sdk.Tx) error {
 	sigs, err := tx.(signing.SigVerifiableTx).GetSignaturesV2()
 	if err != nil {
@@ -94,23 +85,27 @@ func (snm *senderNonceMempool) Insert(_ sdk.Context, tx sdk.Tx) error {
 	nonce := sig.Sequence
 	senderTxs, found := snm.senders[sender]
 	if !found {
-		newSenderTx := newSenderTxs()
-		senderTxs = &newSenderTx
+		senderTxs = huandu.New(huandu.Uint64)
 	}
-	tk := txKey{nonce: nonce, sender: sender}
-	senderTxs.insert(tk, tx)
+	senderTxs.Set(nonce, tx)
 	snm.senders[sender] = senderTxs
 	return nil
 }
 
+// Select returns an iterator ordering transactions the mempool with the lowest nonce of a random selected sender first.
 func (snm *senderNonceMempool) Select(context sdk.Context, i [][]byte) Iterator {
 	var senders []string
+	senderCursors := make(map[string]*senderTxs)
 	for key := range snm.senders {
 		senders = append(senders, key)
+		senderTx := newSenderTxs(snm.senders[key].Front())
+		senderCursors[key] = &senderTx
 	}
+
 	iter := &senderNonceMepoolIterator{
-		mempool: snm,
-		senders: senders,
+		senders:         senders,
+		rnd:             snm.rnd,
+		sendersCurosors: senderCursors,
 	}
 
 	newIter := iter.Next()
@@ -123,8 +118,8 @@ func (snm *senderNonceMempool) Select(context sdk.Context, i [][]byte) Iterator 
 // CountTx returns the total count of txs in the mempool.
 func (snm *senderNonceMempool) CountTx() int {
 	count := 0
-	for senderKey := range snm.senders {
-		count += snm.senders[senderKey].txQueue.Len()
+	for _, value := range snm.senders {
+		count += value.Len()
 	}
 	return count
 }
@@ -143,16 +138,17 @@ func (snm *senderNonceMempool) Remove(tx sdk.Tx) error {
 	sig := sigs[0]
 	sender := sig.PubKey.Address().String()
 	nonce := sig.Sequence
-	tk := txKey{nonce: nonce, sender: sender}
 	senderTxs, found := snm.senders[sender]
 	if !found {
 		return ErrTxNotFound
 	}
-	err = senderTxs.remove(tk)
-	if err != nil {
-		return err
+
+	res := senderTxs.Remove(nonce)
+	if res == nil {
+		return ErrTxNotFound
 	}
-	if senderTxs.txQueue.Len() == 0 {
+
+	if senderTxs.Len() == 0 {
 		delete(snm.senders, sender)
 	} else {
 		snm.senders[sender] = senderTxs
@@ -161,29 +157,31 @@ func (snm *senderNonceMempool) Remove(tx sdk.Tx) error {
 }
 
 type senderNonceMepoolIterator struct {
-	mempool   *senderNonceMempool
-	currentTx *huandu.Element
-	senders   []string
+	rnd             *rand.Rand
+	currentTx       *huandu.Element
+	senders         []string
+	sendersCurosors map[string]*senderTxs
 }
 
 func (i *senderNonceMepoolIterator) Next() Iterator {
 	for len(i.senders) > 0 {
-		senderIndex := i.mempool.rnd.Intn(len(i.senders))
+		senderIndex := i.rnd.Intn(len(i.senders))
 		sender := i.senders[senderIndex]
-		senderTxs, found := i.mempool.senders[sender]
+		senderTxs, found := i.sendersCurosors[sender]
 		if !found {
 			i.senders = removeAtIndex(i.senders, senderIndex)
 			continue
 		}
-		tx := senderTxs.getMove()
+		tx := senderTxs.next()
 		if tx == nil {
 			i.senders = removeAtIndex(i.senders, senderIndex)
 			continue
 		}
 		return &senderNonceMepoolIterator{
-			senders:   i.senders,
-			currentTx: tx,
-			mempool:   i.mempool,
+			senders:         i.senders,
+			currentTx:       tx,
+			rnd:             i.rnd,
+			sendersCurosors: i.sendersCurosors,
 		}
 	}
